# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CashLenX API
  description: Personal finance management API
  version: 0.2.0
  contact:
    name: CashLenX Team
    email: support@cashlenx.com
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/cash/outcome:
    post:
      summary: Create outcome transaction
      description: Record a new expense transaction
      operationId: createOutcome
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashFlowRequest'
      responses:
        '201':
          description: Outcome transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'
  /api/cash/income:
    post:
      summary: Create income transaction
      description: Record a new income transaction
      operationId: createIncome
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashFlowRequest'
      responses:
        '201':
          description: Income transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'
...
components:
  schemas:
...
    ResponseWrapper:
      type: object
      properties:
        data:
          type: object
          description: The actual data payload
        error:
          $ref: '#/components/schemas/ErrorInfo'
        meta:
          $ref: '#/components/schemas/MetaInfo'
      required:
        - data
    ErrorInfo:
      type: object
      properties:
        code:
          type: string
          description: Error code for machine consumption
        message:
          type: string
          description: Human-readable error message
      required:
        - code
        - message
    MetaInfo:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items (for pagination)
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        request_id:
          type: string
          description: Unique request identifier

    # Cash Flow schemas
    CashFlowRequest:
      type: object
      properties:
        belongs_date:
          type: string
          format: date
          description: Transaction date (YYYYMMDD format)
        category_name:
          type: string
          description: Category name for the transaction
        amount:
          type: number
          format: double
          description: Transaction amount
        description:
          type: string
          description: Transaction description
      required:
        - belongs_date
        - category_name
        - amount
...